Swap from fragment into an action in Drawer
