+++ app/src/main/java/com/github/sdp/tarjetakuna/ui/browser/BrowserFragment.kt
@@ -17,8 +17,7 @@ class BrowserFragment : Fragment() {
     // This property is only valid between onCreateView and
     // onDestroyView.
     private val binding get() = _binding!!
-
-
+    
     override fun onCreateView(
         inflater: LayoutInflater,
         container: ViewGroup?,
