+++ app/src/main/res/menu/drawer.xml
@@ -1,8 +1,11 @@
 <?xml version="1.0" encoding="utf-8"?>
 <menu xmlns:android="http://schemas.android.com/apk/res/android"
     xmlns:app="http://schemas.android.com/apk/res-auto">
-    <item android:id="@+id/action_settings"
-        android:title="@string/action_settings"
+    <item
+        android:id="@+id/action_export_collection"
+        android:onClick="onClick"
         android:orderInCategory="100"
+        android:title="@string/menu_exportcollection"
         app:showAsAction="never" />
+    <group android:orderInCategory="101" />
 </menu>
