+++ app/src/main/res/navigation/mobile_navigation.xml
@@ -29,16 +29,10 @@
         android:label="@string/menu_webapi"
         tools:layout="@layout/fragment_web_api" />
 
-	<fragment
+    <fragment
         android:id="@+id/nav_profile"
         android:name="com.github.sdp.tarjetakuna.ui.profile.ProfileFragment"
         android:label="@string/menu_profile"
         tools:layout="@layout/fragment_profile" />
-
-    <fragment
-        android:id="@+id/nav_exportcollection"
-        android:name="com.github.sdp.tarjetakuna.ui.collectionexport.ExportCollectionFragment"
-        android:label="@string/menu_exportcollection"
-        tools:layout="@layout/fragment_export_collection" />
 
 </navigation>
